[AOS] refactor: trade write 패키지 변경
