+++ dodamduck_aos/app/src/main/java/org/chosun/dodamduck/presentation/trade/write/TradeWriteScreen.kt
@@ -1,4 +1,4 @@
-package org.chosun.dodamduck.presentation.trade
+package org.chosun.dodamduck.presentation.trade.write
 
 import android.net.Uri
 import androidx.activity.compose.rememberLauncherForActivityResult
@@ -44,6 +44,7 @@ import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.RequestBody.Companion.toRequestBody
 import org.chosun.dodamduck.R
 import org.chosun.dodamduck.data.model.DodamDuckData
+import org.chosun.dodamduck.presentation.trade.TradeViewModel
 import org.chosun.dodamduck.ui.component.DodamDuckRadioButton
 import org.chosun.dodamduck.ui.component.DodamDuckText
 import org.chosun.dodamduck.ui.component.DodamDuckTextH2
