+++ dodamduck_aos/app/src/main/java/org/chosun/dodamduck/ui/navigation/DodamDuckNavigation.kt
@@ -33,7 +33,7 @@ import org.chosun.dodamduck.presentation.post.detail.PostDetailScreen
 import org.chosun.dodamduck.presentation.post.write.PostWriteScreen
 import org.chosun.dodamduck.presentation.auth.RegisterScreen
 import org.chosun.dodamduck.presentation.search.SearchScreen
-import org.chosun.dodamduck.presentation.trade.TradeWriteScreen
+import org.chosun.dodamduck.presentation.trade.write.TradeWriteScreen
 import org.chosun.dodamduck.ui.UserScreen
 import org.chosun.dodamduck.ui.theme.Brown
